+++ interfaces/plugin/src/sensor_napi_utils.cpp
@@ -272,7 +272,8 @@ void ConvertToSensorData(napi_env env, AsyncCallbackInfo *asyncCallbackInfo, nap
 {
     napi_get_undefined(env, &result[0]);
     int32_t sensorTypeId = asyncCallbackInfo->data.sensorData.sensorTypeId;
-    if ((g_sensorAttributeList.count(sensorTypeId)) == 0 || (g_sensorAttributeList[sensorTypeId].size()
+    if ((g_sensorAttributeList.count(sensorTypeId)) == 0
+        || (static_cast<int32_t>(g_sensorAttributeList[sensorTypeId].size())
         > (asyncCallbackInfo->data.sensorData.dataLength / sizeof(float)))) {
         HiLog::Error(LABEL, "%{public}s count of sensorTypeId is zero", __func__);
         return;
